+++ iftttie/dataclasses_.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import datetime
+from sqlite3 import Row
 from typing import Any, Optional
+from ujson import loads
 
 from iftttie.enums import ValueKind
 
@@ -13,6 +15,15 @@ class Update:
     value: Any = None
     timestamp: datetime = field(default_factory=lambda: datetime.now().astimezone())
     kind: ValueKind = ValueKind.OTHER
+
+    @staticmethod
+    def from_row(row: Row) -> Update:
+        return Update(
+            key=row['key'],
+            value=loads(row['value']),
+            timestamp=datetime.fromtimestamp(row['timestamp'] / 1000).astimezone(),
+            kind=ValueKind(row['kind']),
+        )
 
 
 @dataclass